% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{runMetaboCombiner}
\alias{runMetaboCombiner}
\title{Run metaboCombineR}
\usage{
runMetaboCombiner(listExperimens, windowsize, algorithm, trunc)
}
\arguments{
\item{listExperimens}{a list of data.frames representing experiments}

\item{windowsize}{a number indicating window size parameter for 'rtcor', or k-mer size for 'k-mer' algorithm.}

\item{algorithm}{a string, 'rtcor' or 'kmer' algorithm}

\item{ppm}{a relative value of m/z difference for the peak/feature matching}

\item{abs}{an absolute value of m/z difference for the peak/feature matching}

\item{trunc}{a number of decimal places to be truncated for m/z peak/feature matching}

\item{s_match}{a score for matched features, in an alignment problem (default: 100)}

\item{s_del}{a penalization for deletion of features, in an alignment problem (default: 1)}

\item{s_ins}{a penalization for insertion of features, in an alignment problem (default: 1)}

\item{mzres}{reported mz values reported in pairwise alignment according to reference ("ref"), non-reference ("nonref") experiments, or its average ("avg"). (default: avg)}
}
\value{
Data.frame of final matrix
}
\description{
Run metaboCombineR
}
\details{
Only one parameter of 'ppm', 'abs', or 'trunc' can be specified.
}
\examples{
data(metaboExp1)
data(metaboExp2)
data(metaboExp3)
data(metaboExp4)
runMetaboCombiner(list(metaboExp1, metaboExp2, metaboExp3, metaboExp4), algorithm = "kmer", trunc = 2)
}
