% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{getMZs}
\alias{getMZs}
\title{Get vector of MZ}
\usage{
getMZs(experiment)
}
\arguments{
\item{experiment}{an input data.frame representing experiments}
}
\value{
Vector of MZs
}
\description{
Get vector of MZ
}
\examples{
data(metaboExp1)
getMZs(metaboExp1)
}
