% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{getRTs}
\alias{getRTs}
\title{Get vector of RT}
\usage{
getRTs(experiment)
}
\arguments{
\item{experiment}{an input data.frame representing experiments}
}
\value{
Vector of RTs
}
\description{
Get vector of RT
}
\examples{
data(metaboExp1)
getRTs(metaboExp1)
}
