% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{sortDataFrameByRt}
\alias{sortDataFrameByRt}
\title{Sort experiment according to RT}
\usage{
sortDataFrameByRt(experiment)
}
\arguments{
\item{experiment}{an input data.frame representing experiments}
}
\value{
Vector of MZs
}
\description{
Sort experiment according to RT
}
\examples{
data(metaboExp1)
sortDataFrameByRt(metaboExp1)
}
